Pushing fix to insert tcb
Needs to move head if head prio lower

+++ inc/timeMeasure.h
@@ -2,9 +2,6 @@
 #define __MEASURE_TIME_H
 
 
-
-
-
 void timeMeasureInit(void);  // initialize time measurement. At main, it should be exist
 	
 void timeMeasurestart(void);
@@ -14,7 +11,7 @@ void disableTimeget(void);
 
 void enableTimeget(void);
 
-int getDisablePercent(void);
+float getDisablePercent(void);
